requirements.txt
asgiref==3.8.1
black==24.10.0
click==8.1.7
coverage==7.6.9
Django==5.1.4
djangorestframework==3.15.2
mock==5.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
psycopg==3.2.3
psycopg-binary==3.2.3
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
